requirements.txt
babel==2.17.0
blinker==1.9.0
click==8.3.0
colorama==0.4.6
Flask==3.1.2
greenlet==3.2.4
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mrz==0.6.2
packaging==25.0
psycopg2-binary==2.9.11
SQLAlchemy==2.0.43
tkcalendar==1.6.1
typing_extensions==4.15.0
Werkzeug==3.1.3
